Updated read/write and correlation record (to fix correlation record read glitch)

+++ LIDARLite/examples/Single Sensor/Correlation_Record/Correlation_Record.ino
@@ -10,7 +10,7 @@
 #include <LIDARLite.h>
 
 LIDARLite myLidarLite;
-byte *correlationRecordArray;
+int *correlationRecordArray;
 
 void setup() {
   Serial.begin(115200);
